Basically finished the browserHistoryModule. I also upload a secondary module and an example of client
I upload just the .py files because they are the ones with code. When you fetch for updates you then have to set in into a project and try it with a proper interpreter. This could be a little bit more streamlined in some parts though, we will see if we have time to do that.
In the end, the user's pc runs a Flask app, so that it's reachable from the core of YSDI whenever they want (too frequently doesn't serve any good, maybe 3s is enough). At the beginning, the raspberry posts the blacklist with an http request. In the following queries, it posts the last second where he knew of some visited websites, so that we don't duplicate info.