Streamlined code. Now it's closer to a RESTful API
In the user module I changed the base structure of it all. I changed the functions so that the one sending the blacklist is actually posting, and taking the samples is a get.
In the Raspberry Pi client now I check if the blacklist has been correctly received.
In TimeManagement some functions have been removed since they weren't used at all.