Fixed some little errors that I left during my last commit, removing the microphone from multithreading
Rearranged files in the userPCapp folder, so that you can now open a pycharm project already working from the local repository.
Added the constant RECORD_SECONDS to the transmission that happens from the userPCapp to MultiThreading, since it's used there and I didn't remember yesterday.
Removed the old boardServer.py.
The other files were changed working from pycharm on the so formed project